Docker setup
Docker integration for webserver, node js and redis. Please, while docker is running, navigate to the "docker setup" folder in your terminal and try to execute the command "docker-compose up -d"  This should launch 3 docker containers for redis, for nginx and for nodejs. You can check this using "docker ps". For now as proof of concept i added the index.html to the startpage of localhost meaning You should be able to see the board when going to localhost. Please tell me if there are any problems/errors, the docker setup was quite confusing but it works for me now.