docker setup
docker setup for the graph branch. to setup the docker have docker running and just navigate to the docker setup folder with the terminal and once in the folder use the command "docker-compose up -d" to initialize. This should download all the dependencies and launch a webserver with localhost showing index.html